move graph function to clab package
